# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个基于 Gin 的短链接生成与访问服务",
        "title": "短链接服务 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/",
    "paths": {
        "/api/v1/shorten": {
            "post": {
                "description": "提交原始URL，生成对应的短链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "短链接"
                ],
                "summary": "创建短链接",
                "parameters": [
                    {
                        "description": "创建短链接请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateShortUrlRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "短链接创建成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "自定义短码已存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/s/{shorturl}": {
            "get": {
                "description": "通过短码访问原始URL，自动302重定向",
                "tags": [
                    "短链接"
                ],
                "summary": "访问短链接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "短码",
                        "name": "shorturl",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "重定向到原始URL"
                    },
                    "404": {
                        "description": "短链接不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "410": {
                        "description": "短链接已过期",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CreateShortUrlRequest": {
            "type": "object",
            "properties": {
                "custom_code": {
                    "type": "string",
                    "maxLength": 8,
                    "minLength": 1
                },
                "expire_at": {
                    "type": "string"
                },
                "expire_time": {
                    "type": "string"
                },
                "original_url": {
                    "type": "string"
                }
            }
        }
    }
}